+++ packages/cli/src/commands/multi-chain/deploy.ts
@@ -1,13 +1,14 @@
 // Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
+import assert from 'assert';
 import fs from 'fs';
 import path from 'path';
+import {input} from '@inquirer/prompts';
 import {Command, Flags} from '@oclif/core';
 import {getMultichainManifestPath, getProjectRootAndManifest} from '@subql/common';
 import chalk from 'chalk';
 import cli from 'cli-ux';
-import inquirer from 'inquirer';
 import YAML from 'yaml';
 import {ROOT_API_URL_PROD} from '../../constants';
 import {
@@ -15,15 +16,15 @@ import {
   dictionaryEndpoints,
   executeProjectDeployment,
   generateDeploymentChain,
-  imageVersions,
   ipfsCID_validate,
   processEndpoints,
   projectsInfo,
   splitMultichainDataFields,
 } from '../../controller/deploy-controller';
-import {uploadToIpfs} from '../../controller/publish-controller';
+import {getDirectoryCid, uploadToIpfs} from '../../controller/publish-controller';
 import {MultichainDataFieldType, V3DeploymentIndexerType} from '../../types';
-import {addV, checkToken, promptWithDefaultValues, resolveToAbsolutePath, valueOrPrompt} from '../../utils';
+import {addV, checkToken, resolveToAbsolutePath, valueOrPrompt} from '../../utils';
+import {promptImageVersion} from '../deployment/deploy';
 
 export default class MultiChainDeploy extends Command {
   static description = 'Multi-chain deployment to hosted service';
@@ -70,8 +71,9 @@ export default class MultiChainDeploy extends Command {
     flags.org = await valueOrPrompt(flags.org, 'Enter organisation', 'Organisation is required');
     flags.projectName = await valueOrPrompt(flags.projectName, 'Enter project name', 'Project name is required');
 
-    // Multichain query descriptor
-    const ipfsCID = fileToCidMap.get(path.basename(multichainManifestPath));
+    // Multichain query descriptor, The IPFS provided for deployment here must be a directory
+    const ipfsCID = getDirectoryCid(fileToCidMap);
+    assert(ipfsCID, 'Multichain deployment CID not found');
 
     const projectInfo = await projectsInfo(authToken, flags.org, flags.projectName, ROOT_API_URL_PROD, flags.type);
     const chains: V3DeploymentIndexerType[] = [];
@@ -82,23 +84,13 @@ export default class MultiChainDeploy extends Command {
 
     if (!flags.queryVersion) {
       try {
-        const queryAvailableVersions = await imageVersions(
+        flags.queryVersion = await promptImageVersion(
           multichainManifestObject.query.name,
           multichainManifestObject.query.version,
+          flags.useDefaults,
           authToken,
-          ROOT_API_URL_PROD
+          'Enter query version'
         );
-        if (!flags.useDefaults) {
-          flags.queryVersion = await promptWithDefaultValues(
-            inquirer,
-            `Enter query version`,
-            null,
-            queryAvailableVersions,
-            true
-          );
-        } else {
-          flags.queryVersion = queryAvailableVersions[0];
-        }
       } catch (e) {
         throw new Error(chalk.bgRedBright('Query version is required'));
       }
@@ -114,25 +106,17 @@ export default class MultiChainDeploy extends Command {
         throw new Error(chalk.bgRedBright('Invalid IPFS CID'));
       }
 
+      assert(validator.chainId, 'Please set chainId in your project');
       if (!indexerVersions[validator.chainId]) {
+        assert(validator.manifestRunner, 'Please set manifestRunner in your project');
         try {
-          const indexerAvailableVersions = await imageVersions(
+          indexerVersions[validator.chainId] = await promptImageVersion(
             validator.manifestRunner.node.name,
             validator.manifestRunner.node.version,
+            flags.useDefaults,
             authToken,
-            ROOT_API_URL_PROD
+            `Enter indexer version for ${multichainProjectPath}`
           );
-          if (!flags.useDefaults) {
-            indexerVersions[validator.chainId] = await promptWithDefaultValues(
-              inquirer,
-              `Enter indexer version for ${multichainProjectPath}`,
-              null,
-              indexerAvailableVersions,
-              true
-            );
-          } else {
-            indexerVersions[validator.chainId] = indexerAvailableVersions[0];
-          }
         } catch (e) {
           throw new Error(chalk.bgRedBright('Indexer version is required'), {cause: e});
         }
@@ -149,26 +133,20 @@ export default class MultiChainDeploy extends Command {
           );
         }
 
-        endpoints[validator.chainId] = await promptWithDefaultValues(
-          cli,
-          `Enter endpoint for ${multichainProjectPath}`,
-          undefined,
-          null,
-          true
-        );
+        endpoints[validator.chainId] = await input({
+          message: `Enter endpoint for ${multichainProjectPath}`,
+          required: true,
+        });
       }
 
       if (!dictionaries[validator.chainId]) {
         const validateDictEndpoint = processEndpoints(await dictionaryEndpoints(ROOT_API_URL_PROD), validator.chainId);
         if (!flags.useDefaults && !validateDictEndpoint) {
-          dictionaries[validator.chainId] = await promptWithDefaultValues(
-            cli,
-            `Enter dictionary for ${multichainProjectPath}`,
-            validateDictEndpoint,
-            null,
-            false
-          );
-        } else {
+          dictionaries[validator.chainId] = await input({
+            message: `Enter dictionary for ${multichainProjectPath}`,
+            required: false,
+          });
+        } else if (validateDictEndpoint) {
           dictionaries[validator.chainId] = validateDictEndpoint;
         }
       }
@@ -184,7 +162,7 @@ export default class MultiChainDeploy extends Command {
       );
     }
 
-    this.log('Deploying SubQuery multi-chain project to Hosted Service');
+    this.log('Deploying SubQuery multi-chain project to Hosted Service, IPFS: ', ipfsCID);
 
     await executeProjectDeployment({
       log: this.log.bind(this),