+++ packages/cli/src/controller/project-controller.ts
@@ -17,6 +17,7 @@ export async function createProject(
   organization: string,
   subtitle: string,
   logoUrl: string,
+  project_type: number,
   project_name: string,
   authToken: string,
   gitRepository: string,
@@ -42,11 +43,12 @@ export async function createProject(
         name: project_name,
         subtitle: subtitle,
         dedicateDBKey: dedicateDB,
+        type: project_type,
       },
     });
     return res.data as unknown as CreateProjectResponse;
   } catch (e) {
-    errorHandle(e, 'Failed to create project:');
+    throw errorHandle(e, 'Failed to create project:');
   }
 }
 
@@ -68,6 +70,6 @@ export async function deleteProject(
     });
     return `${key}`;
   } catch (e) {
-    errorHandle(e, 'Failed to delete project:');
+    throw errorHandle(e, 'Failed to delete project:');
   }
 }