+++ packages/node-core/src/indexer/fetch.service.spec.ts
@@ -5,7 +5,16 @@ import {EventEmitter2} from '@nestjs/event-emitter';
 import {SchedulerRegistry} from '@nestjs/schedule';
 import {BaseDataSource, BaseHandler, BaseMapping, DictionaryQueryEntry, IProjectNetworkConfig} from '@subql/types-core';
 import {range} from 'lodash';
-import {BlockDispatcher, delay, IBlock, IBlockDispatcher, IProjectService, NodeConfig} from '../';
+import {
+  BaseUnfinalizedBlocksService,
+  BlockDispatcher,
+  delay,
+  Header,
+  IBlock,
+  IBlockDispatcher,
+  IProjectService,
+  NodeConfig,
+} from '../';
 import {BlockHeightMap} from '../utils/blockHeightMap';
 import {DictionaryService} from './dictionary/dictionary.service';
 import {BaseFetchService} from './fetch.service';
@@ -24,9 +33,6 @@ class TestFetchService extends BaseFetchService<BaseDataSource, IBlockDispatcher
   }
   getGenesisHash(): string {
     return genesisHash;
-  }
-  async getFinalizedHeight(): Promise<number> {
-    return Promise.resolve(this.finalizedHeight);
   }
   async getBestHeight(): Promise<number> {
     return Promise.resolve(this.bestHeight);
@@ -64,6 +70,10 @@ class TestFetchService extends BaseFetchService<BaseDataSource, IBlockDispatcher
   mockDsMap(blockHeightMap: BlockHeightMap<any>): void {
     this.projectService.getDataSourcesMap = jest.fn(() => blockHeightMap);
   }
+
+  async getFinalizedHeader(): Promise<Header> {
+    return Promise.resolve({blockHeight: this.finalizedHeight, blockHash: '0xxx', parentHash: '0xxx'});
+  }
 }
 
 const nodeConfig = new NodeConfig({
@@ -76,7 +86,7 @@ const nodeConfig = new NodeConfig({
 const getNetworkConfig = () =>
   ({
     dictionary: 'https://example.com',
-  } as IProjectNetworkConfig);
+  }) as IProjectNetworkConfig;
 
 const mockDs: BaseDataSource = {
   kind: 'mock/DataSource',
@@ -130,7 +140,7 @@ const getDictionaryService = () =>
     initDictionaries: () => {
       /* TODO */
     },
-  } as any as DictionaryService<any, any>);
+  }) as any as DictionaryService<any, any>;
 
 const getBlockDispatcher = () => {
   const inst = {
@@ -156,6 +166,7 @@ describe('Fetch Service', () => {
   let dictionaryService: DictionaryService<any, any>;
   let networkConfig: IProjectNetworkConfig;
   let dataSources: BaseDataSource[];
+  let unfinalizedBlocksService: BaseUnfinalizedBlocksService<any>;
 
   let spyOnEnqueueSequential: jest.SpyInstance<
     void | Promise<void>,
@@ -199,7 +210,13 @@ describe('Fetch Service', () => {
       blockDispatcher,
       dictionaryService,
       eventEmitter,
-      schedulerRegistry
+      schedulerRegistry,
+      unfinalizedBlocksService,
+      {
+        metadata: {
+          set: jest.fn(),
+        },
+      } as any
     );
 
     spyOnEnqueueSequential = jest.spyOn(fetchService as any, 'enqueueSequential') as any;
@@ -318,7 +335,7 @@ describe('Fetch Service', () => {
   });
 
   it('checks chain heads at an interval', async () => {
-    const finalizedSpy = jest.spyOn(fetchService, 'getFinalizedHeight');
+    const finalizedSpy = jest.spyOn(fetchService, 'getFinalizedHeader');
     const bestSpy = jest.spyOn(fetchService, 'getBestHeight');
 
     await fetchService.init(1);
@@ -332,7 +349,11 @@ describe('Fetch Service', () => {
     expect(finalizedSpy).toHaveBeenCalledTimes(2);
     expect(bestSpy).toHaveBeenCalledTimes(2);
 
-    await expect(fetchService.getFinalizedHeight()).resolves.toBe(fetchService.finalizedHeight);
+    await expect(fetchService.getFinalizedHeader()).resolves.toEqual({
+      blockHeight: fetchService.finalizedHeight,
+      blockHash: '0xxx',
+      parentHash: '0xxx',
+    });
   });
 
   it('enqueues blocks WITHOUT dictionary', async () => {
@@ -655,8 +676,8 @@ describe('Fetch Service', () => {
   });
 
   it('throws if the start block is greater than the chain latest height', async () => {
-    await expect(() => fetchService.init(1001)).rejects.toThrow(
-      `The startBlock of dataSources in your project manifest (1001) is higher than the current chain height (1000). Please adjust your startBlock to be less that the current chain height.`
+    await expect(() => fetchService.init(1002)).rejects.toThrow(
+      `The startBlock of dataSources in your project manifest (1002) is higher than the current chain height (1000). Please adjust your startBlock to be less that the current chain height.`
     );
   });
 
@@ -676,11 +697,31 @@ describe('Fetch Service', () => {
     (fetchService as any).dictionaryService.scopedDictionaryEntries = () => {
       return undefined;
     };
+    fetchService.bestHeight = 500;
     const dictionarySpy = jest.spyOn((fetchService as any).dictionaryService, 'scopedDictionaryEntries');
     await fetchService.init(10);
     expect(dictionarySpy).toHaveBeenCalledTimes(1);
     expect(spyOnEnqueueSequential).toHaveBeenCalledTimes(1);
 
     expect(enqueueBlocksSpy).toHaveBeenLastCalledWith([10, 11, 12, 13, 14, 15, 16, 17, 18, 19], 19);
+  });
+
+  it(`doesn't use dictionary if processing near latest height`, async () => {
+    enableDictionary();
+    (fetchService as any).dictionaryService.scopedDictionaryEntries = () => {
+      return undefined;
+    };
+    fetchService.bestHeight = 500;
+    const dictionarySpy = jest.spyOn((fetchService as any).dictionaryService, 'scopedDictionaryEntries');
+    await fetchService.init(490);
+    expect(dictionarySpy).toHaveBeenCalledTimes(0);
+    expect(spyOnEnqueueSequential).toHaveBeenCalledTimes(1);
+
+    expect(enqueueBlocksSpy).toHaveBeenLastCalledWith([490, 491, 492, 493, 494, 495, 496, 497, 498, 499], 499);
+  });
+
+  it('fetch init when last processed height is same as', async () => {
+    // when last processed height is 1000, finalized height is 1000
+    await expect(fetchService.init(1001)).resolves.not.toThrow();
   });
 });