+++ packages/node-core/src/indexer/poi/poi.service.spec.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import {EventEmitter2} from '@nestjs/event-emitter';
+import {SchedulerRegistry} from '@nestjs/schedule';
 import {Test, TestingModule} from '@nestjs/testing';
 import {delay} from '@subql/common';
 import {Sequelize, Transaction} from '@subql/x-sequelize';
@@ -60,7 +61,7 @@ describe('PoiService', () => {
     } as unknown as NodeConfig;
 
     const sequelize = new Sequelize();
-    storeCache = new StoreCacheService(sequelize, nodeConfig, new EventEmitter2());
+    storeCache = new StoreCacheService(sequelize, nodeConfig, new EventEmitter2(), new SchedulerRegistry());
 
     storeCache.init(true, true, {} as any, {} as any);
     (storeCache as any).cachedModels._metadata = {
@@ -79,6 +80,11 @@ describe('PoiService', () => {
     }).compile();
 
     service = module.get<PoiService>(PoiService);
+  });
+
+  afterEach(async () => {
+    await storeCache?.beforeApplicationShutdown();
+    service.onApplicationShutdown();
   });
 
   it('should be defined', () => {
@@ -125,7 +131,6 @@ describe('PoiService', () => {
     });
 
     it('should not migratePoi if latestSyncedPoiHeight is defined', async () => {
-      storeCache.init(true, true, {} as any, {} as any);
       await service.init('testSchema');
       expect(storeCache.metadata.find).toHaveBeenCalledWith('latestSyncedPoiHeight');
     });